Moved components to delivery
